+++ src/models/configuration.js
@@ -55,10 +55,7 @@ async function writeOAuthConf (oauthData) {
 }
 
 const conf = env.getOrElseAll({
-  API_HOST: 'https://api.clever-cloud.com/v2',
-  // API_HOST: 'https://ccapi-preprod.cleverapps.io/v2',
-  LOG_WS_URL: 'wss://api.clever-cloud.com/v2/logs/logs-socket/<%- appId %>?since=<%- timestamp %>',
-  LOG_HTTP_URL: 'https://api.clever-cloud.com/v2/logs/<%- appId %>',
+  API_HOST: 'https://api.clever-cloud.com',
   EVENT_URL: 'wss://api.clever-cloud.com/v2/events/event-socket',
   WARP_10_EXEC_URL: 'https://c1-warp10-clevercloud-customers.services.clever-cloud.com/api/v0/exec',
   OAUTH_CONSUMER_KEY: 'T5nFjKeHH4AIlEveuGhB5S3xg8T19e',